backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
aiosqlite==0.20.0
cryptography==42.0.7
PyNaCl==1.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pyotp==2.9.0
qrcode[pil]==7.4.2
paramiko==3.4.0
structlog==24.1.0
asyncpg==0.29.0
aiomysql==0.2.0
redis==5.0.3
motor==3.4.0
pymongo==4.6.3
aiohttp==3.9.5
